% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taball.qual.fun.R
\name{taball.qual.fun}
\alias{taball.qual.fun}
\title{Title}
\usage{
taball.qual.fun(data, x_all, x_name_all, nb_dec, na.p = FALSE, titre,
  widths = NULL)
}
\arguments{
\item{widths}{}
}
\description{
Title
}
