% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabcont.qual.fun_.R
\name{tabcont.qual.fun_}
\alias{tabcont.qual.fun_}
\title{Title}
\usage{
tabcont.qual.fun_(data, x_all, x2, xall_name, x2_name, nb_dec,
  pcol = TRUE, plig = TRUE, ptot = TRUE, titre, widths = NULL,
  test = FALSE)
}
\arguments{
\item{test}{}
}
\description{
Title
}
