% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabHR_mult.OS.DSS.R
\name{tabHR_mult.OS.DSS}
\alias{tabHR_mult.OS.DSS}
\title{Title}
\usage{
tabHR_mult.OS.DSS(x_OS, x_DSS, data, a, a2, x_namei, data.OS, data.DSS,
  Split.OS, Split.DSS)
}
\arguments{
\item{Split.DSS}{}
}
\description{
Title
}
