% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabHRall.fun.OS.DSS.R
\name{tabHRall.fun.OS.DSS}
\alias{tabHRall.fun.OS.DSS}
\title{Title}
\usage{
tabHRall.fun.OS.DSS(data, del, x, x2, x_gpe_all, Stratas.OS, Stratas.DSS,
  x_name_all, titre, widths = NULL)
}
\arguments{
\item{widths}{}
}
\description{
Title
}
