% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabHRall.fun_mult.OS.DSS.R
\name{tabHRall.fun_mult.OS.DSS}
\alias{tabHRall.fun_mult.OS.DSS}
\title{Title}
\usage{
tabHRall.fun_mult.OS.DSS(data, Split.OS, Split.DSS, formula.OS,
  formula.DSS, x_var_OS, x_var_DSS, x_name, titre, widths = NULL)
}
\arguments{
\item{widths}{}
}
\description{
Title
}
