% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabHRall.OS.DSS.R
\name{tabHRall.OS.DSS}
\alias{tabHRall.OS.DSS}
\title{Title}
\usage{
tabHRall.OS.DSS(data, del, x, x2, x_gpe_all, x_name_all, Stratas.OS,
  Stratas.DSS, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
