% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabHRall_mult.OS.DSS.R
\name{tabHRall_mult.OS.DSS}
\alias{tabHRall_mult.OS.DSS}
\title{Title}
\usage{
tabHRall_mult.OS.DSS(x_var_OS, x_var_DSS, data, a, a2, x_name, data.OS,
  data.DSS, Split.OS, Split.DSS, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
