% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablogrank.all.report.r
\name{tablogrank.all.report}
\alias{tablogrank.all.report}
\title{Title}
\usage{
tablogrank.all.report(data, x, delai, etat, etat2, nb_dec, titre)
}
\arguments{
\item{titre}{}
}
\description{
Title
}
